+++ CHANGES.md
@@ -1,8 +1,8 @@
 wavesurfer.js changelog
 =======================
 
-4.0.2 (In progress)
--------------------
+Next (unreleased)
+-----------------
 
 - Improved and unified loop playback logic in regions plugin. (#1868)
 - Don't call HTMLMediaElement#load when given peaks and preload == 'none'.