+++ CHANGES.md
@@ -6,6 +6,7 @@ Next (unreleased)
 
 - Use `clip-path` instead of `width` changes in the multicanvcas drawer
   to avoid layout thrashing.
+- Improved and unified loop playback logic in regions plugin. (#1868)
 
 4.0.1 (23.06.2020)
 ------------------