+++ CHANGES.md
@@ -7,6 +7,7 @@ Next (unreleased)
 - Improved and unified loop playback logic in regions plugin. (#1868)
 - Don't call HTMLMediaElement#load when given peaks and preload == 'none'.
   Prevents browsers from pre-fetching audio (#1969)
+- Regions plugin: removed 'col-resize' cursor when resize is disabled (#1985)
 
 4.0.1 (23.06.2020)
 ------------------