+++ CHANGES.md
@@ -5,7 +5,7 @@ Next (unreleased)
 -----------------
 
 - Always call drawBuffer when for media-elements when peaks are provided, fixes
-  a missed call uncovered by #1969
+  a missed call uncovered by #1969 (#1990)
 - Use `clip-path` instead of `width` changes in the multicanvcas drawer
   to avoid layout thrashing. (#1983)
 - Improved and unified loop playback logic in regions plugin. (#1868)