+++ CHANGES.md
@@ -4,13 +4,11 @@ wavesurfer.js changelog
 Next (unreleased)
 -----------------
 
-- Always call drawBuffer when for media-elements when peaks are provided, fixes
-  a missed call uncovered by #1969 (#1990)
 - Use `clip-path` instead of `width` changes in the multicanvcas drawer
   to avoid layout thrashing. (#1983)
 - Improved and unified loop playback logic in regions plugin. (#1868)
 - Don't call HTMLMediaElement#load when given peaks and preload == 'none'.
-  Prevents browsers from pre-fetching audio (#1969)
+  Prevents browsers from pre-fetching audio (#1969, #1990)
 
 4.0.1 (23.06.2020)
 ------------------