+++ CHANGES.md
@@ -11,6 +11,7 @@ Next (unreleased)
 - Regions plugin:
   - Removed 'col-resize' cursor when resize is disabled (#1985)
   - Improved and unified loop playback logic (#1868)
+- Removed `util.extend` method since deprecated (#1995)
 
 4.0.1 (23.06.2020)
 ------------------