+++ CHANGES.md
@@ -9,6 +9,7 @@ Next (unreleased)
 - Regions plugin:
   - Removed 'col-resize' cursor when resize is disabled (#1985)
   - Improved and unified loop playback logic (#1868)
+- Removed util.extend: deprecated since v3.3.0  (#1995)
 
 4.0.1 (23.06.2020)
 ------------------